move ViewType to catalyst
